backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Validation & Serialization
Marshmallow==3.20.1
python-dotenv==1.0.0

# Blockchain & Web3
web3==6.11.2
eth-keys==0.5.1

# Caching
redis==5.0.1

# File Upload
Werkzeug==3.0.1
boto3==1.34.7

# Security
bcrypt==4.1.1
PyJWT==2.8.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Email (Optional)
Flask-Mail==0.9.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production
gunicorn==21.2.0
